changed cfs_ratio division method

+++ proc.h
@@ -54,7 +54,7 @@ struct proc {
   int status;                  // exit status code (assignment 1 task 3)
   long long accumulator;       // for ps_priority scheduling
   int ps_priority;             // for ps_priority scheduling
-  float decay_factor;          // for cfs scheduling
+  double decay_factor;         // for cfs scheduling
   int rtime;                   // running time (cfs scheduling)
   int stime;                   // sleeping time (cfs scheduling)
   int retime;                  // ready time (cfs scheduling)
